requirements.txt
fastapi
# TTS
uvicorn
pydub
PyMuPDF
gtts
google-cloud-texttospeech

pytesseract
PIL